Use ollama without streaming
